+++ packages/core/CHANGELOG.md
@@ -1,5 +1,27 @@
 # fuma-translate
 
+## 1.0.1
+
+### Patch Changes
+
+- 03da3a5: fix CI
+
+## 1.0.0
+
+### Major Changes
+
+- 94f788d: Support JSX syntax in translations
+- 94f788d: Rewrite compiler in Rust
+- 94f788d: Support `fromTranslations` for non-hook usage
+
+### Minor Changes
+
+- bdf0d18: Support CLI usage
+
+### Patch Changes
+
+- 58ff6bf: Support `strict` option
+
 ## 0.0.3
 
 ### Patch Changes